change int to size_t

+++ labs/welcome/code/censor.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 #include <string>
 
-int main(size_t argc, char* argv[]){
+int main(int argc, char* argv[]){
     if(argc != 2){
         std::cout<<"USAGE: censor [length]";
     }
